fix regression: scroll to line, cleanup: highlight

+++ headers/main.h
@@ -17,6 +17,6 @@ extern cchar_t mark;
 extern uint flag; // displayed x processed by offset funcs in size.cpp
 extern uint y, x;
 extern uint maxy, maxx; // to store the maximum rows and columns
-extern long ofy; // offset in y axis of text and screen, x axis is in gapbuffer
+extern uint ofy; // offset in y axis of text and screen, x axis is in gapbuffer
 extern uint ry, rx; // x, y positions in buffer/list
 extern char *filename; // name of open file, if none, it isn't malloc'ed
